refactor conversion bindings in separate source files
